models/01_boar_detector_v2/requirements.txt
ultralytics>=8.3.50
opencv-python-headless
numpy
pillow
folium
